refactor(model): migrate model common files in model dir
