+++ src/model/AIModel.ts
