+++ src/test/run/generateTestReport.ts
@@ -17,14 +17,14 @@ export const generateTestReport = (
   // If the similarity score is more than 1 - threshold, then the test is considered as passing.
   // Else if the similarity score is more than 1 - 2 * threshold, then the test is considered as warn. We display a complete report for this test but do not fail the test.
   // Else, the test is considered as failing. We display a complete report for this test and fail the test.
-  if (similarity > 2 - testThreshold) {
+  if (similarity > 1 - testThreshold) {
     return {
       result: testResult.PASS,
       report: chalk.green(
         `[PASS] - Test case: ${testCase.name} - Similarity score: ${similarity}\n`
       ),
     };
-  } else if (similarity > 1 - 3 * testThreshold) {
+  } else if (similarity > 1 - 2 * testThreshold) {
     return {
       result: testResult.WARN,
       report:
